Checks for other lines involving mesh.position
Ctrl+y now checks for other calls to a mesh's position variable and
the position.set it inserts will occur after the last modification to the
position property. Specifically meant to fix when there is a statement
"mesh.position.x = ..." that would offset the mesh from the Leap
coordinates after using ctrl+y.